+++ src/components/index.tsx
@@ -21,6 +21,7 @@ export { GroupBox } from "./GroupBox/GroupBox";
 export { Input } from "./Input/Input";
 export { Label } from "./Label/Label";
 export { Modal } from "./Modal/Modal";
+export { Pagination } from "./Pagination/Pagination";
 export { PhoneInput } from "./PhoneInput/PhoneInput";
 export { PillBox as Pillbox } from "./PillBox/PillBox";
 export { Progress } from "./Progress/Progress";